refactor(orchestra): extract queue persistence to mixin
Extract queue persistence logic to QueuePersistenceMixin:
- QueueEntry dataclass
- _restore_queue, _persist_queue, get_queued_issue_numbers
- _promote_progressed_entries, _collect_frozen_queue

GlobalDispatchCoordinator reduced from 618 to 415 lines.
